Level Editor compiles now.
Cast a bunch of File Paths to strings. Added constructor for
Common.Geometry.Rectangle  and Common.Geometry.RectangleF  that take
their System.Drawing equivalent. Fixed SetSheetPostion in SpriteEditor.
Fixed args for adding a screen layer. (note the arg had a boolean for if
the layer is a forescreen that I set to false).