Update MainForm.cs
- Trello card: even-if-engine-doesn-t-have-focus-if-pressing-a-key-like-enter-the-engine-will-respond

Problem occured because when program is minimised while clicking on the icon in the system tray, the event that minimizes the form doesn't call the activate/deactivate event from the form load, which are the ones that pause the engine. Same logic applies when clicking on this icon to activate program.

Needed to add code that is sure to be executed when program is minimized/maximized this way.

This function was added: protected override void OnResize(EventArgs e)

Sometime this code will be called to stop engine with OnDeactivate or OnActivated method. This is however the best method I found to be sure engine is activated/deactivated every time it needs to be.