- Engine doesn't stop sometime when it is not visible.
Entity.cs
---------
Function to start/stop engine now checks if engine is already started/stopped before executing code.
It is done because sometimea when it is called status is already ok.

MainForm.cs
-----------
OnResize event added. It stops engine or start it if it needs to be.
If not here and engine is minimised by clicking on tray icon, it won't stop.