- Trello 126 SECOND STEP: Remember options between program usage.  - MainForm.cs (Design): Unchecked many properties of menus. They are now set from MainForm LoadConfig.  - UserSettings.cs: Added classes for serializations  - UserSettings.cs: Added classes for error messages and default values (both for config file reading)  - UserSettingsEnums.cs: Added an enum for NTSC option  - MainForm.cs LoadConfig: Handle loading of parameters. Changed some functions to handle it in a easier way.   - Remember positions of application.
NOTE: Loading of config (LoadConfig) in MainForm.cs MainForm is done later. Else some properties are applied before they can be.
NOTE: In this commit, the write in the settings.xml file is disabled. Will be reactivated in subsequent commits.
NOTE: Do not use this version until Trello 126 is complete for a stable version.