There is always one tile too much shown in editor. It happens as soon as having at least one tile.
To understand the problem, it is needed to explained how observable tiles are shown.

They are shown by this function which reload tiles every time:

        {
            // easier to reassign the whole thing than correctly update the collection
            // wrong I know but don't really care right now.
            _observedTiles = new ObservableCollection<Tile>(_tileset.Tiles);
            OnPropertyChanged("Tiles");
        }

But then, every time a tile is added, this line of code is executed, which add the tiles to observable tiles (however the function mentionned before already display them all, so this is gonna show a tile in excess).

_observedTiles.Add(tile);